Move widget components into subfolder and update imports
