+++ packages/components/src/index.ts
@@ -10,10 +10,13 @@ export * from './BigInt'
 export * from './Markdown'
 export * from './Image'
 export * from './ColorPicker'
-export * from './ReactSelect'
-export * from './CodeEditor'
 export * from './ErrorIndicator'
 
 // The definition factories' override surface; the factory builder itself
 // (`createDefinitionFactory`) stays internal
 export { type DefinitionOverrides } from './_common/createDefinitionFactory'
+
+// Editor-slot widgets (`ReactSelect`, `CodeEditor`) are NOT re-exported here.
+// They replace JsonEditor's built-in UI controls (`Select` / `TextEditor`)
+// rather than render a node type, so they ship under their own subpath,
+// `@json-edit-react/components/widgets` (see ./widgets).
