+++ packages/components/tsconfig.json
@@ -19,5 +19,5 @@
     "noFallthroughCasesInSwitch": true,
     "resolveJsonModule": true
   },
-  "files": ["src/index.ts"]
+  "files": ["src/index.ts", "src/widgets/index.ts"]
 }
